Fixed up the timing. It is running at a 3 MHz clock. The average loop time at the current moment is 500 microseconds roughly. Not too shabby for not trying very hard to optimize what the code. There is a lot of variation from the USB though. I need to tack on the ability to write it to the sdcard as well. Might improve the variance at least.

+++ Firmware/fifo.h
@@ -38,7 +38,7 @@ void   fifo_push_time_event(button_event_t *button); // array of 2
 
 void   fifo_broadcast_state_usb (fsm_state_t current, fsm_state_t next,
                                  fsm_state_t requested, fsm_state_t required,
-                                 unsigned int timing);
+                                 unsigned long int timing);
 bool_t fifo_fetch_usb (usb_data_packet_t *to_be_filled, unsigned char *len);
 void   fifo_push_usb (usb_data_packet_t *to_be_emptied, unsigned char len);
 bool_t fifo_waiting_usb(void); // returns true if there is a message waiting to be processed
